Cleanup for merge
