+++ .github/workflows/code-coverage.yml
@@ -22,12 +22,12 @@ jobs:
         include:
           - job_name: linux-fuse3
             os: ubuntu-latest
-            containerName: 'test-cnt-ubn'
+            containerName: "test-cnt-ubn"
             fuselib: libfuse3-dev
             fuselib2: fuse3
           - job_name: linux-fuse2
             os: ubuntu-latest
-            containerName: 'test-cnt-ubn'
+            containerName: "test-cnt-ubn"
             fuselib: libfuse-dev
             fuselib2: fuse
 
@@ -1037,7 +1037,7 @@ jobs:
         include:
           - job_name: linux
             os: ubuntu-latest
-            containerName: 'test-cnt-ubn'
+            containerName: "test-cnt-ubn"
             fuselib: libfuse3-dev
             fuselib2: fuse3
